requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
plyer>=2.1.0
rich>=13.0.0
aiosqlite>=0.19.0
pytest>=8.0.0
pytest-asyncio>=0.23.0